requirements.txt
# ── Core ──────────────────────────────────────────
redis[hiredis]>=5.0.4      # cliente Redis con hiredis para mejor performance
pydantic>=2.7.0            # validación de esquemas de mensajes (HELLO, INFO, MESSAGE)
anyio>=4.4.0               # compatibilidad con async (base de asyncio/trio)
typer>=0.12.3              # CLI amigable para comandos (run-node, send, etc.)
python-dotenv>=1.0.1

